backend/requirements.txt
Django==4.2
roslibpy==1.5.0
minio==7.1.14
pyMySQL==1.0.3
pylint==2.12.2
pylint-django==2.5.3
pyjwt==2.6.0
PyYAML==6.0
django-cors-headers==3.13.0
numpy==1.24.3
opencv-python-headless
cmake
dlib
face-recognition
